// src/app/shared/model/nft.model.ts
export class NFTModel {
    nftTokenId: number= 0;
    imagePath:string = '';
    name:string = '';
    description: string='';
    price: number = 0;
    isForSale: boolean = true;
    isEquipped: boolean = true;
    category: string = '';
    collection: string = '';
    astroType: string = '';
    rarity: string = '';
    network: string = '';
    blockchainId: string = '';
    collectionId: string = '';
    tokenOwner: string = '';
}